Port to gstreamer 1.0 in a single overly complex patch.
Do not use this for history, I wanted to find all the related changes
and commit as logical patches. Since I found this difficult when finding
what needed to be ported via compiler errors I ended up with a bunch of
changes. However I wanted to share this progress, so I'm commiting this
patch. Hopefully I'll have the time to break it up into smaller
more comprehensible changes.

With this patch, src and the tests build, elements and examples still
don't build.

A few of the test cases even pass.