Replace discovering the video format from caps retrieved from the GstBuffer.
GstBuffers in 1.0 no longer have CAPS attached to them, the paint function
was depending on that to discover what the video format of the buffer was.

So this code has set_caps setting the m_bufferFormat and setting the
BufferFormat dirty flag. The paint function is still responsible for
creating the new painter.