src/QGst/memory.{h,cpp} Provide minimal functional to wrap GstMemory objects.
This just allows creating a GstMemory object, getting its size and
mapping/unmapping it.

The mapping functionality uses a small structure GstMapInfo to hold
information abount and an accessor pointer to the raw memory.

I created a new structure MapInfo which duplicates GstMapInfo because
a goal of QGst was to allow building gstreamer applications without
having gstreamer dev headers installed.

Question: is there a better way of handling that structure?