elements/gstqtvideosink/qtvideosinkdelegate.cpp remove BufferFormat local from paint
Since you can't get caps from a GstBuffer, I had to add the BufferFormat
getter / setters so a message could set the caps and thus
the buffer format.

I removed references to the local buffer format in favor of the
member m_bufferFormat. I did leave the changePainter call in the paint
method to make sure that the change happens in the GUI thread.