elements/gstqtvideosink/autotest.cpp change test ignore reage from 6% to 9%
The test is doing an image comparison between the test pattern being
passed through the rendering pipeline and directly grabbing the image.

For me one of the images always fails with ignoring 6% of the pixels from
the boundary. The last (blue?) color channel is 25 units darker. (230 vs 255).

In addition to whatever changes might have happend in GStreamer I'm assuming the color
blending depends on your screen's resolution, pixel density, and the whim of your
graphics drivers.

The 9% test still fails for me if my screen is cluttered enough that the
different windows get placed on different locations of my screen.